SpeedyExperiments

Documentation for SpeedyExperiments.

SpeedyExperiments.DeviceSetup

SpeedyExperiments.Device

SpeedyExperiments.DeviceArray

SpeedyExperiments.DeviceArray

SpeedyExperiments.DeviceArrayNotAdapt

SpeedyExperiments.DeviceSparseArray

SpeedyExperiments.Device_KA

SpeedyExperiments.gpuoff

SpeedyExperiments.gpuon

SpeedyExperiments.launch_kernel!

SpeedyExperiments.DeviceSetup — Type

DeviceSetup

Holds information about the device the model is running on and workgroup size.

SpeedyExperiments.Device — Method

Device()

Return currently used device for internal purposes, either CPUDevice or GPUDevice

SpeedyExperiments.DeviceArray — Method

DeviceArray(x)

Returns a CuArray when CUDA is used, otherwise a regular Array

SpeedyExperiments.DeviceArray — Method

DeviceArray(device::AbstractDevice, x)

Returns a CuArray when device<:GPUDevice is used, otherwise a regular Array. Uses adapt, thus also can return SubArrays etc.

SpeedyExperiments.DeviceArrayNotAdapt — Method

DeviceArrayNotAdapt(device::AbstractDevice, x)

Returns a CuArray when device<:GPUDevice is used, otherwise a regular Array. Doesn't uses adapt, therefore always returns CuArray/Array

SpeedyExperiments.DeviceSparseArray — Method

DeviceArray(x)

Returns a CUSPARSE Array when CUDA is used, otherwise a regular sparse Array

SpeedyExperiments.Device_KA — Method

Device_KA()

Return currently used device for KernelAbstractions, either CPU or CUDADevice

SpeedyExperiments.gpuoff — Method

gpuoff()

Manually toggle GPU use off

SpeedyExperiments.gpuon — Method

gpuon()

Manually toggle GPU use on (if available)

SpeedyExperiments.launch_kernel! — Method

launch_kernel!(device_setup::DeviceSetup, kernel!, ndrange, kernel_args...)

Launches the kernel! on the device_setup with ndrange computations over the kernel and arguments kernel_args. Returns an event.